feat(apiv2) : add history
